+++ lib/gcm.ex
@@ -1,10 +1,10 @@
 defmodule GCM do
   use Application
 
-  def push(pool, token, alert) do
+  def push(pool, token, notification) do
     msg = GCM.Message.new
     |> Map.put(:token, token)
-    |> Map.put(:alert, alert)
+    |> Map.put(:notification, notification)
     push(pool, msg)
   end