+++ lib/gcm/worker.ex
@@ -158,8 +158,8 @@ defmodule GCM.Worker do
       payload = put_in(payload[:aps][:sound], msg.sound)
     end
 
-    if msg.category != nil do
-      payload = put_in(payload[:aps][:category], msg.category)
+    if msg.topic != nil do
+      payload = put_in(payload[:aps][:topic], msg.topic)
     end
 
     if msg.badge != nil do
@@ -170,14 +170,14 @@ defmodule GCM.Worker do
       payload = put_in(payload[:aps][:'content-available'], msg.content_available)
     end
 
-    if msg.extra != [] do
-      payload = Map.merge(payload, msg.extra)
+    if msg.data != [] do
+      payload = Map.merge(payload, msg.data)
     end
 
-    if is_binary(msg.alert) do
-      payload = put_in(payload[:aps][:alert], msg.alert)
+    if is_binary(msg.notification) do
+      payload = put_in(payload[:aps][:notification], msg.notification)
     else
-      payload = put_in(payload[:aps][:alert], format_loc(msg.alert))
+      payload = put_in(payload[:aps][:notification], format_loc(msg.notification))
     end
 
     encode(payload, payload_limit)
@@ -187,22 +187,22 @@ defmodule GCM.Worker do
     json = Poison.encode!(payload)
 
     length_diff = byte_size(json) - payload_limit
-    length_alert = case payload.aps.alert do
+    length_notification = case payload.aps.notification do
       %{body: body} -> byte_size(body)
       str when is_binary(str) -> byte_size(str)
     end
 
     cond do
       length_diff <= 0 -> json
-      length_diff >= length_alert -> {:error, :payload_size_exceeded}
+      length_diff >= length_notification -> {:error, :payload_size_exceeded}
       true ->
-        payload = put_in(payload[:aps][:alert], truncate(payload.aps.alert, length_alert - length_diff))
+        payload = put_in(payload[:aps][:notification], truncate(payload.aps.notification, length_notification - length_diff))
         Poison.encode!(payload)
     end
   end
 
-  defp truncate(%{body: string} = alert, size) do
-    %{alert | body: truncate(string, size)}
+  defp truncate(%{body: string} = notification, size) do
+    %{notification | body: truncate(string, size)}
   end
 
   defp truncate(string, size) when is_binary(string) do
@@ -217,28 +217,28 @@ defmodule GCM.Worker do
 
   defp format_loc(%GCM.Message.Loc{title: title, body: body, title_loc_key: title_loc_key,
                                     title_loc_args: title_loc_args, action_loc_key: action_loc_key,
-                                    loc_key: loc_key, loc_args: loc_args,
+                                    body_loc_key: body_loc_key, body_loc_args: body_loc_args,
                                     launch_image: launch_image}) do
     # These are required parameters
-    alert = %{title: title, body: body, "loc-key": loc_key, "loc-args": loc_args}
+    notification = %{title: title, body: body, "loc-key": body_loc_key, "loc-args": body_loc_args}
     # Following are optional parameters
     if title_loc_key != nil do
-      alert = alert
+      notification = notification
       |> Map.put(:'title-loc-key', title_loc_key)
     end
     if title_loc_args != nil do
-      alert = alert
+      notification = notification
       |> Map.put(:'title-loc-args', title_loc_args)
     end
     if action_loc_key != nil do
-      alert = alert
+      notification = notification
       |> Map.put(:'action-loc-key', action_loc_key)
     end
     if launch_image != nil do
-      alert = alert
+      notification = notification
       |> Map.put(:'launch-image', launch_image)
     end
-    alert
+    notification
   end
 
   defp send_message(socket, msg, payload) do