+++ test/apns_test.exs
@@ -11,10 +11,10 @@ defmodule GCMTest do
     end
   end
 
-  test "If PN exceeds length of 256 bytes, it still builds proper payload by truncating alert message" do
+  test "If PN exceeds length of 256 bytes, it still builds proper payload by truncating notification message" do
     msg = %GCM.Message{}
     |> Map.put(:token, String.duplicate("0", 64))
-    |> Map.put(:alert, String.duplicate("lorem ipsum", 100))
+    |> Map.put(:notification, String.duplicate("lorem ipsum", 100))
     payload = GCM.Worker.build_payload(msg, 256)
     assert byte_size(payload) == 256
     assert payload =~ "…"
@@ -23,7 +23,7 @@ defmodule GCMTest do
   test "PN with UTF8-characters is properly truncated" do
     msg = %GCM.Message{}
     |> Map.put(:token, String.duplicate("0", 64))
-    |> Map.put(:alert, String.duplicate("ありがとう", 30))
+    |> Map.put(:notification, String.duplicate("ありがとう", 30))
     payload = GCM.Worker.build_payload(msg, 256)
     # When truncating UTF8 chars, payload size may be less than 256
     assert byte_size(payload) <= 256
@@ -34,7 +34,7 @@ defmodule GCMTest do
     string = "lorem ipsum"
     msg = %GCM.Message{}
     |> Map.put(:token, String.duplicate("0", 64))
-    |> Map.put(:alert, string)
+    |> Map.put(:notification, string)
     payload = GCM.Worker.build_payload(msg, 256)
     assert byte_size(payload) == @payload_min_size + byte_size(string)
     refute payload =~ "…"
@@ -44,7 +44,7 @@ defmodule GCMTest do
     string = String.duplicate("a", 256 - @payload_min_size)
     msg = %GCM.Message{}
     |> Map.put(:token, String.duplicate("0", 64))
-    |> Map.put(:alert, string)
+    |> Map.put(:notification, string)
     payload = GCM.Worker.build_payload(msg, 256)
     assert byte_size(payload) == @payload_min_size + byte_size(string)
     refute payload =~ "…"
@@ -54,7 +54,7 @@ defmodule GCMTest do
     string = "test123 тест テスト !@#$%"
     msg = %GCM.Message{}
     |> Map.put(:token, String.duplicate("0", 64))
-    |> Map.put(:alert, string)
+    |> Map.put(:notification, string)
     payload = GCM.Worker.build_payload(msg, 256)
     assert byte_size(payload) == @payload_min_size + byte_size(string)
     refute payload =~ "…"